Version 2.2 - Centrality and More Stats!
Updated the Java Parser to output more information about the network in
the JSON files. Network graphs now display centrality (finally!). Colors
are not consistant however, but most of the times central nodes are
depicted in Orange, and other nodes are depicted in Blue. More stats are
now displayed along with the density and enhaced the overall display
experience for home.html.